add pgarray equals override fix issue #3170
