+++ pgjdbc/src/test/java/org/postgresql/test/jdbc2/ArrayTest.java
@@ -20,6 +20,7 @@ import org.postgresql.util.PSQLException;
 
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
@@ -902,6 +903,40 @@ public class ArrayTest extends BaseTest4 {
             ars.getMetaData().getColumnType(1));
       }
     }
+  }
+
+  @Test
+  public void testArrayEquals() throws SQLException {
+    Statement statement =  conn.createStatement();
+
+    statement.executeQuery("SELECT * FROM person");
+    Array pgArray1 = new PgArray((BaseConnection) conn, Oid.BYTEA_ARRAY, new byte[]{'1','2','3'});
+    Array pgArray2 = new PgArray((BaseConnection) conn, Oid.BYTEA_ARRAY, new byte[]{'1','2','3'});
+    Assertions.assertEquals(pgArray1,pgArray2);
+
+    Array pgArray3 = new PgArray((BaseConnection) conn, Oid.BYTEA_ARRAY, new byte[]{1,2,3});
+    Array pgArray4 = new PgArray((BaseConnection) conn, Oid.BYTEA_ARRAY, new byte[]{1,2,3});
+    Assertions.assertEquals(pgArray3,pgArray4);
+
+    Array pgArray5 = new PgArray((BaseConnection) conn, Oid.BIT_ARRAY, new byte[]{1,2,3});
+    Array pgArray6 = new PgArray((BaseConnection) conn, Oid.BYTEA_ARRAY, new byte[]{1,2,3});
+    Assertions.assertNotEquals(pgArray5,pgArray6);
+
+    Array pgArray7 = new PgArray((BaseConnection) conn, Oid.JSON, new byte[]{1,2,3});
+    Array pgArray8 = new PgArray((BaseConnection) conn, Oid.BYTEA_ARRAY, new byte[]{1,2,3});
+    Assertions.assertNotEquals(pgArray7,pgArray8);
+
+    Array pgArray9 = new PgArray((BaseConnection) conn, Oid.VARCHAR, "{}");
+    Array pgArray10 = new PgArray((BaseConnection) conn, Oid.VARCHAR, "{}");
+    Assertions.assertEquals(pgArray9,pgArray10);
+
+    Array pgArray11 = new PgArray((BaseConnection) conn, Oid.VARCHAR, "{\t \n 'testing1', \t \n 'testing2'}");
+    Array pgArray12 = new PgArray((BaseConnection) conn, Oid.VARCHAR, "{\t \n 'testing1', \t \n 'testing2'}");
+    Assertions.assertEquals(pgArray11,pgArray12);
+
+    Array pgArray13 = new PgArray((BaseConnection) conn, Oid.VARCHAR_ARRAY, "{\t \n 'testing1', \t \n 'testing2'}");
+    Array pgArray14 = new PgArray((BaseConnection) conn, Oid.VARCHAR, "{\t \n 'testing1', \t \n 'testing2'}");
+    Assertions.assertNotEquals(pgArray13,pgArray14);
   }
 
 }
